+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueUi.kt
@@ -1,0 +1,11 @@
+package org.simple.clinic.home.overdue
+
+interface OverdueUi {
+  fun showOverdueAppointments(
+      pendingAppointments: List<OverdueAppointment>,
+      agreedToVisitAppointments: List<OverdueAppointment>,
+      remindToCallLaterAppointments: List<OverdueAppointment>,
+      removedFromOverdueAppointments: List<OverdueAppointment>,
+      moreThanAnYearOverdueAppointments: List<OverdueAppointment>
+  )
+}