+++ app/src/test/java/org/simple/clinic/home/overdue/OverdueUiRendererTest.kt
@@ -1,0 +1,46 @@
+package org.simple.clinic.home.overdue
+
+import com.nhaarman.mockitokotlin2.mock
+import com.nhaarman.mockitokotlin2.verify
+import org.junit.Test
+import org.simple.sharedTestCode.TestData
+import java.util.UUID
+
+class OverdueUiRendererTest {
+
+  private val ui = mock<OverdueUi>()
+  private val uiRenderer = OverdueUiRenderer(
+      ui = ui,
+      isOverdueSectionsFeatureEnabled = true
+  )
+  private val defaultModel = OverdueModel.create()
+
+  @Test
+  fun `when overdue appointments are loaded and overdue sections is enabled, then show overdue appointments`() {
+    // given
+    val pendingAppointments = listOf(TestData.overdueAppointment(appointmentUuid = UUID.fromString("b9c7b7f5-a9e4-4589-9cb9-5b92f650d7b0")))
+    val agreedToVisitAppointments = listOf(TestData.overdueAppointment(appointmentUuid = UUID.fromString("9cb24c2a-02f9-4eec-aa05-d06ba4fcae82")))
+    val removedFromOverdueAppointments = listOf(TestData.overdueAppointment(appointmentUuid = UUID.fromString("f2220d4d-96f7-4f23-9b2f-8f14e59a09df")))
+    val overdueAppointmentsLoadedModel = defaultModel
+        .currentFacilityLoaded(TestData.facility(uuid = UUID.fromString("b5e72d35-73e2-444b-a266-a02b73a6299a")))
+        .overdueAppointmentsLoaded(
+            pendingAppointments = pendingAppointments,
+            agreedToVisitAppointments = agreedToVisitAppointments,
+            remindToCallLaterAppointments = emptyList(),
+            removedFromOverdueAppointments = removedFromOverdueAppointments,
+            moreThanAnYearOverdueAppointments = emptyList()
+        )
+
+    // when
+    uiRenderer.render(overdueAppointmentsLoadedModel)
+
+    // then
+    verify(ui).showOverdueAppointments(
+        pendingAppointments = pendingAppointments,
+        agreedToVisitAppointments = agreedToVisitAppointments,
+        remindToCallLaterAppointments = emptyList(),
+        removedFromOverdueAppointments = removedFromOverdueAppointments,
+        moreThanAnYearOverdueAppointments = emptyList()
+    )
+  }
+}